Initial Split Into Modules, Bump PR Relating To Window Restore Fixes
