+++ click_helpers.py
@@ -1,0 +1,49 @@
+import time
+import pyautogui as auto
+import python_imagesearch.imagesearch as imagesearch
+
+from screen_helpers import onscreen
+import generic_helpers
+
+def click_key(key, delay=.1):
+    auto.keyDown(key)
+    time.sleep(delay)
+    auto.keyUp(key)
+
+
+def click_left(delay=.1):
+    auto.mouseDown()
+    time.sleep(delay)
+    auto.mouseUp()
+
+
+def click_right(delay=.1):
+    auto.mouseDown(button='right')
+    time.sleep(delay)
+    auto.mouseUp(button='right')
+
+
+def click_to(path, precision=0.8, delay=.1):
+    if onscreen(path, precision):
+        auto.moveTo(imagesearch.imagesearch(path))
+        click_left(delay)
+    else:
+        print(f"Could not find '{path}', skipping")
+
+def click_to_multiple(images, conditional_func=None, delay=None):
+    for image in images:
+        try:
+            click_to(image)
+        except Exception:
+            print(f"Failed to click {image}")
+        if generic_helpers.is_var_number(delay):
+            time.sleep(delay)
+        if generic_helpers.is_var_function(conditional_func) and conditional_func():
+            return True
+    return False
+
+def search_to(path):
+    pos = imagesearch.imagesearch(path)
+    if onscreen(path):
+        auto.moveTo(pos)
+        return pos
